Separate DraweeController creation and initialization
Summary:
Currently, DraweeController creation & initialization are mixed, which duplicates quite some code and makes it impossible to create controller-dependant components that are passed to the controller when initialized (i.e. listeners).

NOTE: This is a breaking change for classes that extend `PipelineDraweeController`. You need to adjust your constructor to match the new version. If you are manually creating controllers, make sure to properly initialize them.

Reviewed By: foghina

Differential Revision: D7672781

fbshipit-source-id: 806fc0d0be2d75f4638754a86e7b1fd57ea1b408